[KYUUBI #2021] Command OptionParser for launching Hive Backend Engine
